trabajo final. Oficializar resultado de partida (juez). Manejo de oficializacion de finales.

+++ backend-y-poo/trabajo_final/Trabajo_Final/Trabajo_Final/Controllers/TorneosController.cs
@@ -236,7 +236,7 @@ namespace Trabajo_Final.Controllers
                 id_juez,
                 (int) dto.id_partida,
                 (int) dto.id_ganador,
-                (int) dto.id_descalificado
+                dto.id_descalificado
                 );
 
             return Ok(new { message = $"La partida {dto.id_partida} se oficializó con éxito." });
